third_party/blink/web_tests/svg/dynamic-updates/SVGLinearGradientElement-svgdom-y1-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'y1' property of the SVGLinearGradientElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS linearGradientElement.y1.baseVal.valueAsString is "0%"
PASS linearGradientElement.y1.baseVal.valueAsString is "1000%"
PASS successfullyParsed is true

TEST COMPLETE
